refactor: 크루 별 출석 저장소 생성 및 조회 기능
